Added install script for apt-get packages.  Fixed a data narrowing warning.

+++ src/Bathy.cpp
@@ -60,7 +60,7 @@ Grid* simulatetopographyGrid(int XDist, int YDist) {
 }
 
 
-Grid* getBathy(string inputFile, string inputFileType, int startX, int startY, int XDist, int YDist, string seriesName, long timestamp) {
+Grid* getBathy(string inputFile, string inputFileType, size_t startX, size_t startY, size_t XDist, size_t YDist, string seriesName, long timestamp) {
 	Grid* topographyGrid = new Grid(XDist, YDist, "Topography");
 
    // This will be the netCDF ID for the file and data variable.
